+++ src/managers/btc_only/multi_sig_wallet_manager.c
@@ -34,7 +34,8 @@ static int deleteNode(char *verifyCode);
 static int getLength();
 static void traverseList(void (*callback)(MultiSigWalletItem_t *, void *), void *any);
 static void destroyMultiSigWalletList();
-static void saveToFlash(const char *password);
+static void saveToFlash(void);
+static int findAvailableIndex(void);
 static void modifyNode(char *verifyCode, MultiSigWalletItem_t *newItem);
 static MultiSigWalletItem_t *findNode(char *verifyCode);
 static void DestoryMultisigWalletManager(MultiSigWalletManager_t *manager);
@@ -54,6 +55,7 @@ MultiSigWalletManager_t *initMultiSigWalletManager()
     g_multisigWalletManager->saveToFlash = saveToFlash;
     g_multisigWalletManager->modifyNode = modifyNode;
     g_multisigWalletManager->findNode = findNode;
+    g_multisigWalletManager->findAvailableIndex = findAvailableIndex;
     g_multisigWalletManager->list = NULL;
     createMultiSigWalletList();
     return g_multisigWalletManager;
@@ -77,19 +79,44 @@ MultiSigWalletItem_t *GetMultisigWalletByVerifyCode(const char *verifyCode)
     return g_multisigWalletManager->findNode((char *)verifyCode);
 }
 
-int DeleteMultisigWalletByVerifyCode(const char *verifyCode, const char *password)
+int DeleteMultisigWalletByVerifyCode(const char *verifyCode)
 {
     ASSERT_WALLET_MANAGER_EXIST
     int index = g_multisigWalletManager->deleteNode((char *)verifyCode);
-    g_multisigWalletManager->saveToFlash(password);
+    g_multisigWalletManager->saveToFlash();
     return index;
 }
 
-int GetCurrentAccountMultisigWalletNum(void)
+int GetCurrentAccountMultisigWalletNum(bool isPassphrase)
 {
     ASSERT_WALLET_MANAGER_EXIST
-    g_multisigWalletManager->getLength();
-    return getLength();
+    int length = g_multisigWalletManager->getLength();
+    int passPhraseNum = 0;
+    for (int i = 0; i < length; i++) {
+        if (GetCurrenMultisigWalletByIndex(i)->passphrase == true) {
+            passPhraseNum++;
+            break;
+        }
+    }
+    if (isPassphrase) {
+        return passPhraseNum;
+    }
+    return length - passPhraseNum;
+}
+
+MultiSigWalletItem_t *GetCurrenMultisigWalletByOrder(int index)
+{
+    ASSERT_WALLET_MANAGER_EXIST
+    MultiSigWalletList_t *list = g_multisigWalletManager->list;
+    MultiSigWalletNode_t *temp = list->head;
+
+    while (temp != NULL) {
+        if (temp->value->order == index) {
+            return temp->value;
+        }
+        temp = temp->next;
+    }
+    return NULL;
 }
 
 MultiSigWalletItem_t *GetCurrenMultisigWalletByIndex(int index)
@@ -115,13 +142,11 @@ MultiSigWalletItem_t *GetDefaultMultisigWallet(void)
     CURRENT_WALLET_INDEX_ENUM defaultIndex = GetCurrentWalletIndex();
     MultiSigWalletList_t *list = g_multisigWalletManager->list;
     MultiSigWalletNode_t *temp = list->head;
-    int innerIndex = 0;
     while (temp != NULL) {
-        if (innerIndex == defaultIndex) {
+        if (temp->value->order == defaultIndex) {
             return temp->value;
         }
         temp = temp->next;
-        innerIndex++;
     }
     return NULL;
 }
@@ -149,10 +174,13 @@ MultiSigWalletItem_t *AddMultisigWalletToCurrentAccount(MultiSigWallet *wallet, 
 
     walletItem->network = wallet->network;
 
-    walletItem->order = manager->getLength(manager->list);
+    walletItem->order = manager->findAvailableIndex();
 
     walletItem->verifyCode = MULTI_SIG_MALLOC(MAX_VERIFY_CODE_LENGTH);
     strcpy_s(walletItem->verifyCode, MAX_VERIFY_CODE_LENGTH, wallet->verify_code);
+
+    walletItem->verifyWithoutMfp = MULTI_SIG_MALLOC(MAX_VERIFY_CODE_LENGTH);
+    strcpy_s(walletItem->verifyWithoutMfp, MAX_VERIFY_CODE_LENGTH, wallet->verify_without_mfp);
 
     walletItem->format = MULTI_SIG_MALLOC(MAX_FORMAT_LENGTH);
     strcpy_s(walletItem->format, MAX_FORMAT_LENGTH, wallet->format);
@@ -160,8 +188,10 @@ MultiSigWalletItem_t *AddMultisigWalletToCurrentAccount(MultiSigWallet *wallet, 
     walletItem->walletConfig = MULTI_SIG_MALLOC(MAX_WALLET_CONFIG_TEXT_LENGTH);
     strcpy_s(walletItem->walletConfig, MAX_WALLET_CONFIG_TEXT_LENGTH, wallet->config_text);
 
+    walletItem->passphrase = PassphraseExist(GetCurrentAccountIndex());
+
     manager->insertNode(walletItem);
-    manager->saveToFlash(password);
+    manager->saveToFlash();
     return walletItem;
 }
 
@@ -220,6 +250,35 @@ static void modifyNode(char *verifyCode, MultiSigWalletItem_t *newItem)
         }
         temp = temp->next;
     }
+}
+
+static int findAvailableIndex(void)
+{
+    ASSERT_WALLET_MANAGER_EXIST
+    MultiSigWalletList_t *list = g_multisigWalletManager->list;
+
+    if (list->head == NULL) {
+        return 0;
+    }
+
+    for (int index = 0; index < MAX_MULTI_SIG_WALLET_NUMBER; index++) {
+        bool found = false;
+        MultiSigWalletNode_t *temp = list->head;
+
+        while (temp != NULL) {
+            if (temp->value->order == index) {
+                found = true;
+                break;
+            }
+            temp = temp->next;
+        }
+
+        if (!found) {
+            return index;
+        }
+    }
+
+    return MAX_MULTI_SIG_WALLET_NUMBER;
 }
 
 static MultiSigWalletItem_t *findNode(char *verifyCode)
@@ -317,8 +376,8 @@ static void destroyMultiSigWalletList()
     manager->list = NULL;
 }
 
-static void saveToFlash(const char *password)
+static void saveToFlash(void)
 {
     // todo  判断是否需要更新
-    MultiSigWalletSave(password, g_multisigWalletManager);
+    MultiSigWalletSave(g_multisigWalletManager);
 }